+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec2.0.0.html).
 
+
+## v1.2.8 - Dec 22, 2025
+
+### Added
+
+- Add getGoodImageUrlFromType function in utils.ts
+- Add storage.ts for option page
+- Add data logic to option page
+
+### Changed
+
+- Update utils.ts
+
 ## v1.2.7 - Dec 21, 2025
 
 ### Added